Refactor and update main/views module contents, documentation and view logic
Update the internal logic of view classes in main/views where it is outdated. Add missing docstrings. Move non-main-specific views (such us FielValidationView, BaCa2LoggedInView) and mixins to util/views.
