+++ BaCa2/BaCa2/urls.py
@@ -1,33 +1,23 @@
-"""BaCa2 URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/4.1/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
 from django.contrib import admin
 from django.urls import path, include
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
 
 from main.views import BaCa2LoginView, BaCa2LogoutView
+from util.views import FieldValidationView
 
 urlpatterns = [
-    # path('polls/', include('polls.urls')),
     path('baca/', admin.site.urls),
 
-    # necessary for broker api
-    path('broker_api/', include('broker_api.urls')),
+    # ------------------------------------- Authentication ------------------------------------- #
     path('login/', BaCa2LoginView.as_view(), name='login'),
     path('logout/', BaCa2LogoutView.as_view(), name='logout'),
+
+    # ------------------------------------------ Apps ------------------------------------------ #
+    path('broker_api/', include('broker_api.urls')),
     path('main/', include('main.urls')),
+
+    # --------------------------------------- Auxiliary ---------------------------------------- #
+    path('field_validation', FieldValidationView.as_view(), name='field-validation'),
 ]
 
 urlpatterns += staticfiles_urlpatterns()
