+++ BaCa2/templates/widget_templates/forms/field.html
@@ -63,7 +63,7 @@
                     this,
                     '{{ form_widget.form_cls }}',
                     '{{ form_widget.field_min_length|get_item:field.name }}',
-                    '{% url 'main:field-validation' %}');
+                    '{% url 'field-validation' %}');
             });
         });
     </script>
